Deduplicate Dav1dTaskContext
