+++ src/decode.rs
@@ -441,38 +441,9 @@ use crate::src::levels::BlockSize;
 use crate::src::levels::BS_128x128;
 use crate::src::levels::BS_4x4;
 use crate::src::levels::BS_64x64;
-#[derive(Copy, Clone)]
-#[repr(C)]
-pub struct Dav1dTaskContext {
-    pub c: *const Dav1dContext,
-    pub f: *const Dav1dFrameContext,
-    pub ts: *mut Dav1dTileState,
-    pub bx: libc::c_int,
-    pub by: libc::c_int,
-    pub l: BlockContext,
-    pub a: *mut BlockContext,
-    pub rt: refmvs_tile,
-    pub c2rust_unnamed: Dav1dTaskContext_cf,
-    pub al_pal: [[[[uint16_t; 8]; 3]; 32]; 2],
-    pub pal_sz_uv: [[uint8_t; 32]; 2],
-    pub txtp_map: [uint8_t; 1024],
-    pub scratch: Dav1dTaskContext_scratch,
-    pub warpmv: Dav1dWarpedMotionParams,
-    pub lf_mask: *mut Av1Filter,
-    pub top_pre_cdef_toggle: libc::c_int,
-    pub cur_sb_cdef_idx_ptr: *mut int8_t,
-    pub tl_4x4_filter: Filter2d,
-    pub frame_thread: Dav1dTaskContext_frame_thread,
-    pub task_thread: Dav1dTaskContext_task_thread,
-}
-use crate::src::internal::Dav1dTaskContext_frame_thread;
-use crate::src::internal::Dav1dTaskContext_task_thread;
+use crate::src::internal::Dav1dTaskContext;
 use crate::src::levels::Filter2d;
-
 use crate::src::levels::FILTER_2D_BILINEAR;
-
-use crate::src::internal::Dav1dTaskContext_cf;
-use crate::src::internal::Dav1dTaskContext_scratch;
 use crate::src::refmvs::refmvs_tile;
 
 #[derive(Copy, Clone)]
@@ -2228,6 +2199,7 @@ unsafe fn obmc_lowest_px(
     }
 }
 
+#[deny(unsafe_op_in_unsafe_fn)]
 unsafe fn decode_b(
     t: &mut Dav1dTaskContext,
     bl: BlockLevel,
@@ -6145,7 +6117,7 @@ unsafe extern "C" fn decode_sb(
             (*(node as *const EdgeBranch)).split[0],
         );
     }
-    let mut pc = &mut Default::default();
+    let mut pc = Default::default().as_mut_slice();
     let mut bp: BlockPartition = PARTITION_NONE;
     let mut ctx = 0;
     let mut bx8 = 0;
