+++ src/log.rs
@@ -10,9 +10,6 @@ extern "C" {
 }
 use crate::include::stdatomic::atomic_int;
 use crate::src::r#ref::Dav1dRef;
-
-use crate::include::dav1d::headers::Dav1dWarpedMotionParams;
-
 use crate::include::dav1d::headers::Dav1dSequenceHeader;
 
 use crate::include::dav1d::headers::Dav1dFilmGrainData;
@@ -390,40 +387,9 @@ use crate::include::stdatomic::atomic_uint;
 use crate::src::cdf::CdfThreadContext;
 use crate::src::internal::Dav1dTileGroup;
 use crate::src::picture::Dav1dThreadPicture;
-#[derive(Copy, Clone)]
-#[repr(C)]
-pub struct Dav1dTaskContext {
-    pub c: *const Dav1dContext,
-    pub f: *const Dav1dFrameContext,
-    pub ts: *mut Dav1dTileState,
-    pub bx: libc::c_int,
-    pub by: libc::c_int,
-    pub l: BlockContext,
-    pub a: *mut BlockContext,
-    pub rt: refmvs_tile,
-    pub c2rust_unnamed: Dav1dTaskContext_cf,
-    pub al_pal: [[[[uint16_t; 8]; 3]; 32]; 2],
-    pub pal_sz_uv: [[uint8_t; 32]; 2],
-    pub txtp_map: [uint8_t; 1024],
-    pub scratch: Dav1dTaskContext_scratch,
-    pub warpmv: Dav1dWarpedMotionParams,
-    pub lf_mask: *mut Av1Filter,
-    pub top_pre_cdef_toggle: libc::c_int,
-    pub cur_sb_cdef_idx_ptr: *mut int8_t,
-    pub tl_4x4_filter: Filter2d,
-    pub frame_thread: Dav1dTaskContext_frame_thread,
-    pub task_thread: Dav1dTaskContext_task_thread,
-}
-use crate::src::internal::Dav1dTaskContext_frame_thread;
-use crate::src::internal::Dav1dTaskContext_task_thread;
+use crate::src::internal::Dav1dTaskContext;
 use crate::src::internal::FrameTileThreadData;
-use crate::src::levels::Filter2d;
-
-use crate::src::internal::Dav1dTaskContext_cf;
-use crate::src::internal::Dav1dTaskContext_scratch;
 use crate::src::lf_mask::Av1Filter;
-use crate::src::refmvs::refmvs_tile;
-
 use crate::src::env::BlockContext;
 use crate::src::refmvs::refmvs_frame;
 use crate::src::refmvs::refmvs_temporal_block;
